+++ ng2Client/app/modules/data/entities/UserResourcePool.ts
@@ -1,4 +1,6 @@
-﻿export class UserResourcePool {
+﻿import { EntityBase } from "./entity-base";
+
+export class UserResourcePool extends EntityBase {
 
     // Server-side
     UserId: number = 0;